pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "synthdoc"
version = "0.1.0"
description = "Synthetic document generation for training document understanding models"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "faker>=20.0",
    "pillow>=10.0",
    "opencv-python-headless>=4.8",
    "numpy>=1.26",
    "matplotlib>=3.8",
    "weasyprint>=60.0",
    "jinja2>=3.1",
    "sympy>=1.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
    "mypy>=1.7",
]

[project.scripts]
synthdoc = "synthdoc.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/synthdoc"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
